+++ src/Tests/NativeInteropIntegrationTests.cs
@@ -10,6 +10,7 @@ using DINOForge.SDK;
 using DINOForge.SDK.Assets;
 using DINOForge.SDK.Dependencies;
 using DINOForge.SDK.NativeInterop;
+using DINOForge.Runtime.Assets;
 using FluentAssertions;
 using Xunit;